refactor(frontend-manage): create separate session creation component
