+++ apps/frontend-manage/src/components/sessions/creation/SessionCreation.tsx
@@ -1,0 +1,102 @@
+import { useQuery } from '@apollo/client'
+import { Course, GetUserCoursesDocument } from '@klicker-uzh/graphql/dist/ops'
+import {
+  Tab,
+  TabContent,
+  TabList,
+  Tabs,
+  ThemeContext,
+} from '@uzh-bf/design-system'
+import { useContext, useMemo } from 'react'
+import { twMerge } from 'tailwind-merge'
+import LearningElementCreationForm from '../../../components/sessions/creation/LearningElementCreationForm'
+import LiveSessionCreationForm from '../../../components/sessions/creation/LiveSessionCreationForm'
+import MicroSessionCreationForm from '../../../components/sessions/creation/MicroSessionCreationForm'
+
+function SessionCreation() {
+  const theme = useContext(ThemeContext)
+
+  const {
+    loading: loadingCourses,
+    error: errorCourses,
+    data: dataCourses,
+  } = useQuery(GetUserCoursesDocument)
+
+  const courseSelection = useMemo(
+    () =>
+      dataCourses?.userCourses?.map((course: Partial<Course>) => ({
+        label: course.displayName || '',
+        value: course.id || '',
+      })),
+    [dataCourses]
+  )
+
+  return (
+    <Tabs defaultValue="live-session">
+      <TabList
+        className={{
+          root: 'flex flex-row justify-between w-full h-8 border-b border-solid border-uzh-grey-60',
+        }}
+      >
+        <Tab
+          key="live-session"
+          value="live-session"
+          label="Live-Session"
+          className={{
+            root: twMerge('flex-1', theme.primaryBgHover),
+            label: 'font-bold text-base flex flex-col justify-center h-full',
+          }}
+        />
+        <div className="border-r-2 border-solid border-uzh-grey-60" />
+        <Tab
+          key="micro-session"
+          value="micro-session"
+          label="Micro-Session"
+          className={{
+            root: twMerge(
+              'flex-1 disabled:text-uzh-grey-80 disabled:hover:bg-white disabled:cursor-not-allowed',
+              theme.primaryBgHover
+            ),
+            label: 'font-bold text-base flex flex-col justify-center h-full',
+          }}
+        />
+        <div className="border-r-2 border-solid border-uzh-grey-60" />
+        <Tab
+          key="learning-element"
+          value="learning-element"
+          label="Lernelement"
+          className={{
+            root: twMerge(
+              'flex-1 disabled:text-uzh-grey-80 disabled:hover:bg-white disabled:cursor-not-allowed',
+              theme.primaryBgHover
+            ),
+            label: 'font-bold text-base flex flex-col justify-center h-full',
+          }}
+        />
+      </TabList>
+      <TabContent
+        key="live-session"
+        value="live-session"
+        className={{ root: 'overflow-y-scroll md:h-72' }}
+      >
+        <LiveSessionCreationForm courses={courseSelection} />
+      </TabContent>
+      <TabContent
+        key="micro-session"
+        value="micro-session"
+        className={{ root: 'overflow-y-scroll md:h-72' }}
+      >
+        <MicroSessionCreationForm />
+      </TabContent>
+      <TabContent
+        key="learning-element"
+        value="learning-element"
+        className={{ root: 'overflow-y-scroll md:h-72' }}
+      >
+        <LearningElementCreationForm />
+      </TabContent>
+    </Tabs>
+  )
+}
+
+export default SessionCreation
